change dots to radiating waves. replaced <circle> with <use>
Using CSS animation for the radiating waves. There are several styles
collaborating to create this visual effect.

@keyframes       define stop points changing circle radius and opacity
.wave circle     declare looping and a linear interpolation between keyframes
var(--dt)        delegates the animation start time to individual circles
var(--waveStep)  declares space between the waves

Changed from <circle> to <use> to allow a group of circles.
<defs>           defines the template of a group of circles
<use>            reference the group of circles with a single DOM element

That single DOM element makes the change to all the javascript
easier. Instead of getting and setting the cx and cy attributes of a
<circle> we can get and set the x and y attributes of a <use>. No
change was needed for code that toggles the visibility attribute.